meme-lizard/requirements.txt
Pillow>=9.5
pyttsx3>=2.90
pygame>=2.5
